refactor(setup): rename packages
