+++ src/main/java/me/clementino/holiday/ApiHolidayApplication.java
@@ -1,4 +1,4 @@
-package me.clementino.api_holiday;
+package me.clementino.holiday;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
