+++ src/main/java/me/clementino/holiday/controller/HomeController.java
@@ -1,4 +1,4 @@
-package me.clementino.api_holiday.controller;
+package me.clementino.holiday.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
